Fixed bugs in the ourvoice component, added the Home component, and made the ourvoice component more dynamic
