+++ src/components/Sections/Home/Home.js
@@ -1,0 +1,11 @@
+import Ourvoice from './Ourvoice/Ourvoice';
+
+function Home() {
+  return (
+    <div>
+      <Ourvoice />
+    </div>
+  );
+}
+
+export default Home;
