+++ src/components/Sections/Home/Ourvoice/Card.json
@@ -1,0 +1,23 @@
+[
+  {
+    "title": "Let's talk about react",
+    "content": "A short tour of react to a technology that changed my life and continues to change the industry.",
+    "readers": "300 readers",
+    "information-links": "Read a little more",
+    "id": 1
+  },
+  {
+    "title": "Elegan code in react",
+    "content": "There are many ways to write code in react, but do you know how to do it elegantly?",
+    "readers": "300 readers",
+    "information-links": "Read a little more",
+    "id": 2
+  },
+  {
+    "title": "Tips from a junior",
+    "content": "Some tips from a junior frontend developer that you may not have known and probably need.",
+    "readers": "300 readers",
+    "information-links": "Read a little more",
+    "id": 3
+  }
+]
