+++ src/components/Sections/Home/Ourvoice/ContainerRight.js
@@ -1,0 +1,7 @@
+import Cards from './Cards';
+
+function ContainerRight() {
+  return <Cards />;
+}
+
+export default ContainerRight;
